fix: Route,GroupRoute 엔티티 삭제 후, Group 엔티티로 연관관계 수정
