+++ src/main/java/com/ebiz/wsb/domain/guardian/dto/GuardianDTO.java
@@ -19,5 +19,4 @@ public class GuardianDTO {
     private String experience;
     private String phone;
     private String imagePath;
-    private Long routeId;
 }
