--- src/main/java/com/ebiz/wsb/domain/route/api/RouteController.java
@@ -1,39 +1,0 @@
-package com.ebiz.wsb.domain.route.api;
-
-import com.ebiz.wsb.domain.route.application.RouteService;
-import com.ebiz.wsb.domain.route.dto.RouteDTO;
-import com.ebiz.wsb.domain.waypoint.dto.WaypointDTO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/routes")
-@RequiredArgsConstructor
-public class RouteController {
-
-    private final RouteService routeService;
-
-//    @GetMapping
-//    public ResponseEntity<List<RouteDTO>> getAllRoutes(){
-//        List<RouteDTO> routes = routeService.getAllRoutes();
-//        return ResponseEntity.ok(routes);
-//    }
-//
-//    @GetMapping("/{routeId}")
-//    public ResponseEntity<RouteDTO> getRouteById(@PathVariable Long routeId){
-//        RouteDTO routeDTO = routeService.getRouteById(routeId);
-//        return ResponseEntity.ok(routeDTO);
-//    }
-
-    @GetMapping("/waypoints")
-    public ResponseEntity<List<WaypointDTO>> getWaypoints() {
-        List<WaypointDTO> waypoints = routeService.getWaypoints();
-        return ResponseEntity.ok(waypoints);
-    }
-}
