--- src/main/java/com/ebiz/wsb/domain/route/dto/GroupRouteAssignRequest.java
@@ -1,12 +1,0 @@
-package com.ebiz.wsb.domain.route.dto;
-
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class GroupRouteAssignRequest {
-    private Long routeId;
-    private Long groupId;
-}
