--- src/main/java/com/ebiz/wsb/domain/route/entity/GroupRoute.java
@@ -1,32 +1,0 @@
-package com.ebiz.wsb.domain.route.entity;
-
-import com.ebiz.wsb.domain.group.entity.Group;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicInsert;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@DynamicInsert
-@Table(name = "GroupRoute")
-public class GroupRoute {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "group_id")
-    private Group group;
-
-    @ManyToOne
-    @JoinColumn(name = "route_id")
-    private Route route;
-}
