--- src/main/java/com/ebiz/wsb/domain/route/entity/Route.java
@@ -1,37 +1,0 @@
-package com.ebiz.wsb.domain.route.entity;
-
-import com.ebiz.wsb.domain.guardian.entity.Guardian;
-import com.ebiz.wsb.domain.waypoint.entity.Waypoint;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicInsert;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@DynamicInsert
-@Table(name = "Route")
-public class Route {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "route_id")
-    private Long id;
-
-    @Column(name = "route_name")
-    private String routeName;
-
-    @OneToMany(mappedBy = "route")
-    private List<Waypoint> waypoints = new ArrayList<>();
-
-    @OneToMany(mappedBy = "route")
-    private List<Guardian> guardians = new ArrayList<>();
-}
