--- src/main/java/com/ebiz/wsb/domain/route/exception/RouteNotFoundException.java
@@ -1,12 +1,0 @@
-package com.ebiz.wsb.domain.route.exception;
-
-public class RouteNotFoundException extends RuntimeException{
-
-    public RouteNotFoundException(String message){
-        super(message);
-    }
-
-    public RouteNotFoundException(String message, Throwable cause){
-        super(message, cause);
-    }
-}
