--- src/main/java/com/ebiz/wsb/domain/route/repository/RouteRepository.java
@@ -1,9 +1,0 @@
-package com.ebiz.wsb.domain.route.repository;
-
-import com.ebiz.wsb.domain.route.entity.Route;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RouteRepository extends JpaRepository<Route, Long> {
-}
