+++ src/main/java/com/ebiz/wsb/domain/waypoint/api/WaypointController.java
@@ -2,6 +2,7 @@ package com.ebiz.wsb.domain.waypoint.api;
 
 import com.ebiz.wsb.domain.student.dto.StudentDTO;
 import com.ebiz.wsb.domain.waypoint.application.WaypointService;
+import com.ebiz.wsb.domain.waypoint.dto.WaypointDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +18,12 @@ import java.util.List;
 public class WaypointController {
 
     private final WaypointService waypointService;
+
+    @GetMapping("/")
+    public ResponseEntity<List<WaypointDTO>> getWaypoints() {
+        List<WaypointDTO> waypoints = waypointService.getWaypoints();
+        return ResponseEntity.ok(waypoints);
+    }
 
     @GetMapping("/{waypointId}/students")
     public ResponseEntity<List<StudentDTO>> getStudentsByWaypoint(@PathVariable Long waypointId) {
